[IE TESTS] Move specific import_export_tests to gna and myriad
